requirements.txt
anthropic>=0.40.0
weasyprint>=62.0
jinja2>=3.1.0
python-dotenv>=1.0.0
questionary>=2.0.0
